FW-863. Commit with partial implementation of the 6LoWPAN fragmentation feature.
(1) Copy address information from original message to fragments

(2) Bug fixes and further integration of 6LoWPAN fragmentation.
    - fix checksum calculation of UPD and TCP (in case message length > 128)
    - allow uecho to allocate OpenQueueBigPackets
    - remove some useless code from openbridge
    - update packetfunctions to handle big packets (> 127 bytes)
    - update the uecho.py script to be more interactive

+++ openstack/cross-layers/packetfunctions.h
@@ -36,10 +36,10 @@ void     packetfunctions_readAddress(uint8_t* payload, uint8_t type, open_addr_t
 void     packetfunctions_writeAddress(OpenQueueEntry_t* msg, open_addr_t* address, bool littleEndian);
 
 // reserving/tossing headers and footers
-void     packetfunctions_reserveHeaderSize(OpenQueueEntry_t* pkt, uint8_t header_length);
-void     packetfunctions_tossHeader(OpenQueueEntry_t* pkt, uint8_t header_length);
-void     packetfunctions_reserveFooterSize(OpenQueueEntry_t* pkt, uint8_t header_length);
-void     packetfunctions_tossFooter(OpenQueueEntry_t* pkt, uint8_t header_length);
+void     packetfunctions_reserveHeaderSize(OpenQueueEntry_t* pkt, uint16_t header_length);
+void     packetfunctions_tossHeader(OpenQueueEntry_t* pkt, uint16_t header_length);
+void     packetfunctions_reserveFooterSize(OpenQueueEntry_t* pkt, uint16_t header_length);
+void     packetfunctions_tossFooter(OpenQueueEntry_t* pkt, uint16_t header_length);
 
 // packet duplication
 void packetfunctions_duplicatePacket(OpenQueueEntry_t* dst, OpenQueueEntry_t* src);
